Shutdown audio properly for three types of exit
If audio does not shutdown properly, the audio might be malfunction
after multiple launching of SDL application.

The SDL Quit event of Quake does not shutdown audio properly, so
newquakerootfs.cpio is the tweaked version which calling H_shutdown()
when handle the SDL Quit event(check the latest-buildroot/buildroot).

TODO: rephrase all the comments of this commit.